refactor: Separate fixtures between data and configs
