+++ app/configurator/configurator-state.spec.tsx
@@ -23,11 +23,11 @@ import {
 } from "@/configurator/configurator-state";
 import { DimensionMetaDataFragment } from "@/graphql/query-hooks";
 import { DataCubeMetadata } from "@/graphql/types";
-import bathingWaterMetadata from "@/test/__fixtures/api/DataCubeMetadataWithComponentValues-bathingWater.json";
-import covid19Metadata from "@/test/__fixtures/api/DataCubeMetadataWithComponentValues-covid19.json";
-import covid19ColumnChartConfig from "@/test/__fixtures/dev/chartConfig-column-covid19.json";
-import covid19TableChartConfig from "@/test/__fixtures/dev/chartConfig-table-covid19.json";
-import { data as fakeVizFixture } from "@/test/__fixtures/prod/line-1.json";
+import covid19ColumnChartConfig from "@/test/__fixtures/config/dev/chartConfig-column-covid19.json";
+import covid19TableChartConfig from "@/test/__fixtures/config/dev/chartConfig-table-covid19.json";
+import { data as fakeVizFixture } from "@/test/__fixtures/config/prod/line-1.json";
+import bathingWaterMetadata from "@/test/__fixtures/data/DataCubeMetadataWithComponentValues-bathingWater.json";
+import covid19Metadata from "@/test/__fixtures/data/DataCubeMetadataWithComponentValues-covid19.json";
 
 const mockedApi = api as jest.Mocked<typeof api>;
