+++ app/test/__fixtures/config/dev/chartConfig-column-covid19.json
