+++ app/test/__fixtures/config/dev/chartConfig-table-covid19.json
