+++ app/test/__fixtures/config/int/column-waldflasche.json
