+++ app/test/__fixtures/config/int/scatterplot-palmer-penguins.json
