+++ app/test/__fixtures/data/DataCubeMetadataWithComponentValues-bathingWater.json
