+++ app/test/__fixtures/data/DataCubeMetadataWithComponentValues-covid19.json
