+++ cypress/integration/chart-snapshots.spec.ts
@@ -1,4 +1,4 @@
-import intConfigs from "../../app/test/__fixtures/int/configs";
+import intConfigs from "../../app/test/__fixtures/config/int/configs";
 
 // Right now the CI app server runs connected to int.lindas.admin.ch
 const configs = intConfigs.map((x) => ({ env: "int", ...x }));
